backend/requirements.txt
# Install Dependencies with "pip install -r requirements.txt"

flask==2.3.2
google-cloud-speech==2.15.1
librosa==0.10.0
numpy==1.26.4
python-dotenv==0.21.0
opensmile==2.5.1
joblib
scikit-learn
pandas
torch
einops
torchvision
matplotlib
transformers
garminconnect
google-genai
transformers
